backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
websockets>=12.0
aiofiles>=23.2.1
mutagen>=1.47.0
# Production
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
